P2/P2/requirements.txt
django==4.0.3
pillow==9.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
cryptography~=36.0.2
cffi~=1.15.0
pip~=21.2.4
py~=1.11.0
lxml~=4.8.0
pytz~=2021.3
sqlparse~=0.4.2
asgiref~=3.5.0
colorama~=0.4.4
Jinja2~=2.11.3
setuptools~=58.1.0
requests~=2.27.1
PyJWT~=2.3.0
accounts~=0.2.4